+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLClaimsQueryIdentifier.kt
@@ -1,6 +1,6 @@
 package at.asitplus.openid.dcql
 
-import at.asitplus.data.validation.third_party.kotlin.String.requireIsNotEmpty
+import at.asitplus.data.validation.third_party.kotlin.requireIsNotEmpty
 import kotlinx.serialization.Serializable
 import kotlin.jvm.JvmInline
 
@@ -13,15 +13,11 @@ import kotlin.jvm.JvmInline
 @JvmInline
 value class DCQLClaimsQueryIdentifier(val string: String) {
     init {
-        validate()
-    }
-
-    private fun validate() {
         string.requireIsNotEmpty()
-        string.forEach {
-            if(it != '_' && it != '-' && !it.isLetterOrDigit()) {
-                throw IllegalArgumentException("Value must only consist of alphanumeric, underscore (_) or hyphen (-) characters.")
-            }
+        require(string.all {
+            it == '_' || it == '-' || it.isLetterOrDigit()
+        }) {
+            "Claims query identifier must only contain alphanumeric, underscore (_) or hyphen (-) characters."
         }
     }
 }
