+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLCredentialQueryIdentifier.kt
@@ -1,6 +1,6 @@
 package at.asitplus.openid.dcql
 
-import at.asitplus.data.validation.third_party.kotlin.String.requireIsNotEmpty
+import at.asitplus.data.validation.third_party.kotlin.requireIsNotEmpty
 import kotlinx.serialization.Serializable
 import kotlin.jvm.JvmInline
 
@@ -14,15 +14,11 @@ import kotlin.jvm.JvmInline
 @JvmInline
 value class DCQLCredentialQueryIdentifier(val string: String) {
     init {
-        validate()
-    }
-
-    fun validate() {
         string.requireIsNotEmpty()
-        string.forEach {
-            if(it != '_' && it != '-' && !it.isLetterOrDigit()) {
-                throw IllegalArgumentException("Value must only consist of alphanumeric, underscore (_) or hyphen (-) characters.")
-            }
+        require(string.all {
+            it == '_' || it == '-' || it.isLetterOrDigit()
+        }) {
+            "Credential query identifier must only contain alphanumeric, underscore (_) or hyphen (-) characters."
         }
     }
 }
